requirements.txt
python-dotenv>=1.0.0
python-telegram-bot[job-queue]>=20.0
langgraph
langchain
langchain-core
langchain-openai
openai
httpx<0.28.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
supabase>=2.3.0
python-dateutil==2.9.0
cryptography==42.0.0
psycopg2-binary==2.9.10
langsmith>=0.1.0
fastapi>=0.100.0
uvicorn>=0.30.0

# LangGraph Native Postgres Checkpointer
langgraph-checkpoint-postgres
psycopg[binary]
psycopg-pool
cachetools>=5.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Observability (OpenTelemetry via LangSmith)
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-logging
langsmith[otel]
